git unification
This version is where I am while merging the different parts of this project into one unified git repo. The PCB is at v2, the version first sent to be manufactured. The 3D model is at a point where it should be compatible with the PCB, and various issues with printing it have been addressed, but not tested. The spi2midi stuff used to be in its own repo; the PCB and 3D model I just kept in a succession of dated/versioned directories. I've done my best to represent the history up to this point in this repo.